rm old screenshot stuff
